requirements.txt
anyio==4.9.0
APScheduler==3.11.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
et_xmlfile==2.0.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gspread==6.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
oauthlib==3.3.1
openpyxl==3.1.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dotenv==1.1.0
python-telegram-bot==22.1
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
sniffio==1.3.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0

